Wait for online players before connecting to server console
This should prevent situations where the server will briefly have an online_ping followed by another status without online_ping (which can happen during server startup). If the client registers event handlers on the first connect and sends delayed messages, there is a risk of sending messages over a closed connection because the second status would disconnect the console. Adding a player count check should prevent this because now even when online_ping is received, we do not connect to the server console until an online player is confirmed, making it far less likely that the status will bounce.